+++ model_navigator/runners/tensorrt.py
@@ -22,7 +22,10 @@ from typing import List, Optional, Sequence
 import numpy as np
 
 from model_navigator.api.config import Format, TensorType
-from model_navigator.configuration.validation.device import get_id_from_device_string, validate_device_string_for_cuda
+from model_navigator.configuration.validation.device import (
+    get_id_from_device_string,
+    validate_device_string_for_cuda,
+)
 from model_navigator.core.logger import LOGGER
 from model_navigator.core.tensor import get_tensor_type
 from model_navigator.exceptions import ModelNavigatorError, ModelNavigatorUserInputError
@@ -36,7 +39,6 @@ from model_navigator.utils import common as utils
 from model_navigator.utils import module
 
 trt = module.lazy_import("tensorrt")
-torch = module.lazy_import("torch")
 torch = module.lazy_import("torch")